Try a different singularity image
CRAB jobs produced an error saying:

"No setuid installation found, for unprivileged installation use:
./mconfig --without-suid"

If I'm not mistaken, then the CRAB jobs themselves are sandboxed in a
container (singularity or docker) which can be a problem. Not sure if
trying a different image actually does anything because the issue may be
with the "host container" that runs the actual job. Grasping at straws
at the moment.